# apidog/elloot-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Elloot API",
    "description": "Marketplace API (escrow). Import this file in Apidog: Import → OpenAPI.\n\nAuth: use `Authorization: Bearer <accessToken>` from register/login.\nSandbox flow: create order → checkout → POST /payments/sandbox/confirm → deliver → confirm.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local (.env.local PORT=5000)"
    },
    {
      "url": "http://localhost:3333",
      "description": "Local default (.env.example)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT accessToken from /api/auth/register or /api/auth/login"
      }
    },
    "schemas": {
...
      "SandboxConfirmBody": {
        "type": "object",
        "required": ["providerRef"],
        "properties": {
          "providerRef": {
            "type": "string",
            "description": "From checkout.providerRef (e.g. sandbox_<orderId>)",
            "example": "sandbox_cmsXXXX"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/payments/sandbox/confirm": {
      "post": {
        "tags": ["Payments"],
        "summary": "Sandbox: simulate payment paid (buyer JWT)",
        "operationId": "sandboxConfirm",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SandboxConfirmBody" }
            }
          }
        },
        "responses": {
          "200": { "description": "Payment confirmed" }
        }
      }
    },
    "/api/payments/sandbox/webhook": {
      "post": {
        "tags": ["Payments"],
        "summary": "Sandbox webhook (no JWT, service context)",
        "operationId": "sandboxWebhook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SandboxConfirmBody" }
            }
          }
        },
        "responses": {
          "200": { "description": "Webhook received" }
        }
      }
    },
    "/api/payments/sandbox/confirm-admin": {
      "post": {
        "tags": ["Payments"],
        "summary": "Sandbox confirm as ADMIN",
        "operationId": "sandboxConfirmAdmin",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SandboxConfirmBody" }
            }
          }
        },
        "responses": {
          "200": { "description": "Confirmed" },
          "403": { "description": "Admin only" }
        }
      }
    },
...
  }
}